Renamed API route from `.../parts/#/...` to `.../part/#/...`, and added a HTTP test case for message response endpoint
