+++ src/pages/api/lesson/[lessonId]/part/[partNumber]/index.ts
