+++ test/http/post_response.http
@@ -1,0 +1,6 @@
+POST http://localhost:8080/api/lesson/1/part/0/response HTTP/1.1
+content-type: application/json
+
+{
+	"message": "This is a test message"
+}
